traffic route algoritmn

+++ localizaai/app/src/main/java/com/localizaai/Model/TrafficRequest.kt
@@ -11,6 +11,6 @@ data class TrafficResponse (
     val freeFlowSpeed : Int,
     val currentTravelTime : Int,
     val freeFlowTravelTime : Int,
-    val confidence : Int,
+    val confidence : Double,
     val roadClosure : Boolean
 )
